build a `authentication` module (#37)
